Removes hard dependency on activerecord
