+++ lib/degu/init_active_record_plugins.rb
@@ -1,0 +1,7 @@
+require 'degu/has_set'
+require 'degu/has_enum'
+
+class ActiveRecord::Base
+  include Degu::HasEnum
+  include Degu::HasSet
+end
