--- lib/degu/polite.rb
@@ -1,6 +1,0 @@
-module Degu
-end
-require 'degu/version'
-require 'degu/renum'
-require 'degu/has_set'
-require 'degu/has_enum'
