+++ spec/spec_helper.rb
@@ -1,3 +1,5 @@
 require 'rspec'
 
 require 'degu'
+require 'active_record'
+require 'degu/init_active_record_plugins'
